requirements-dev.txt
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
tox>=4.0.0
ruff>=0.1.0
mypy>=1.0.0
pre-commit>=3.0.0
